Fix issue with where ansible couldn't change permissions on needed files
This fix is not ideal. It's insecure and causes warnings during
provisioning. However, it seems after the upgrade to ubuntu 20.04 this
issue started happening. It appears from the docs that using 'become' in
the ansible task definition is not working. Perhaps read the
documentation in the below error that occurs when this is not in place.
However, I believe I am already doing this. I believe we need to
'become' scott so that permissions are properly set. Perhaps there is a
better way to handle all of this.

fatal: [new]: FAILED! => {"msg": "Failed to set permissions on the
temporary files Ansible needs to create when becoming an unprivileged
user (rc: 1, err: chown: changing ownership of
'/var/tmp/ansible-tmp-1591651136.4772868-4502750366308/': Operation not
permitted\nchown: changing ownership of
'/var/tmp/ansible-tmp-1591651136.4772868-4502750366308/AnsiballZ_stat.py':
Operation not permitted\n}). For information on working around this, see
https://docs.ansible.com/ansible/become.html#becoming-an-unprivileged-user"}