Set up two provisioning profiles for Windows and Mac hosts
There are differences in how we want to provision the Linux vm
depending on if we are on a Windows or Mac host. Currently the
only difference here is if we want to install vbox guest additions
since we use parallels on mac. However, there may be more changes
in the future.

Unfortunately right now this means that we are going to have to
update both files if we want to make a change. Since this is not
ideal we should look into using ansible features such as roles to
allow for these to be combined back into 1 playbook and add some
conditional logic using roles or whatever is appropriate.