Fix issue where ansible didn't have access to files it needed
Ansible was using setfacl in order to access files it needed during
provisioning. Before a hack was used in the ansible.cfg in order to
work around this but it wasn't ideal. This seems like a cleaner fix.
It seems like these machines should have this tool installed on them
or at very least ansible expects them to. See the below github issue
for more information.

https://github.com/ansible/ansible/issues/16052#issuecomment-399038452