Don't automatically update dotfiles
If this is turned on then any changes to tracked dotfiles are clobbered.
Sometimes we might want to make changest to local dotfiles and either
test them for a while or only keep them for the local machine. If this
is on it would revert all changes to dotfiles to the state of the remote
repo when provisioning happened