Move install of oh-my-zsh out of the package install file
This is an install specific to the scott user so it needs to happen
after they are created.

Additionally, it might be a good idea to think about reordering where
these task files are run. For instance, perhaps the package
installations could wait until the end.

- Don't become_user scott when installing terraform. It shouldn't be
necessary since it's a system wide install