Don't copy id_rsa to machine as it no longer exists in the repo
The file was a secret and sensitive and shouldn't have been in source
control. We'll have to either just not do this or allow for it to be
provided at runtime.

Note that if it isn't provided we should skip this task because when
testing testing this in the build pipeline we won't be able to specify
it easily. We could pass it in as a param but that seems rather too
involved and kinda ugly?