Run apt update if first one fails
For some reason after the update it seems like certain packages aren't
available, particularly python and acl. They are available after an
update when I try it myself so I'm not sure what the issue is. This fix
might not work since it's only running update once if the update is
successful (which it still seems to be with the previous method). If
this doesn't work I will just run the update twice regardless. I
actually maybe remember reading something about this where running
update twice is needed for some reason.