Bring build project into this project
Having two separate project for the build and the application was
becoming too cumbersome. They are tightly coupled so it makes sense for
them to be in the same repo. It was extremely common for a change in one
to also require a change in the other. It also makes more sense from a
design standpoint.