Copy credential files into place
This is continuing with the slightly hacky behavior mentioned in the
TODO above in the code. However, for the playbook to work these files
are required to be in place.

It might be a good idea to consider these files "optional". I would have
to look into how to skip steps if the files don't exist but it shouldn't
be hard. These files can be optional (at least in the current state of
the playbook) because nothing in the playbook depends on them. Unlike
the id_rsa key which git clone's depend on. However, if we wanted to
test that, for example, awscli was installed correctly then we'd need
the aws credential file.