Centralize user file copying
There was a lot of repetition in how we were copying files into the
users home directory. Also due to the differences in what files we copy
to desktop and server roles this caused this copying to be split up into
multiple locations. This change centralizes where the copying takes
place (in the common role) and what files are copied is configured in
the role variables.