Remove over-usage of become and become_user
I believe become isn't necessary because at the start of the playbook
it's set for the entire playbook. become_user is still necessary when
running some commands which are doing things in the users home
directory. I've still reduced usages of these as well by using blocks
which run all commands in them using a specific user.