Start VM on boot
This adds a systemd service that will start the VM on boot. I believe it
will also shut down the VM on shutdown, however the vboxmanage command
that's used doesn't actually wait for the VM to stop. So we'd have to
add some kind of wait.

The HAOS VM can also now be managed (start and stop) via systemctl.