Install VBox extension pack
This seemed to be needed for VBox to discover our USB devices.

This was spurred from a Bluetooth USB dongle not being visible in the VM
or by VBox in general even though it was visible to the host. It seems
that installing this is what allowed it to be discovered by VBox.

On systems where BT is build in this may not be necessary but it seems
to supply some useful functionality that may be useful in other
situations so I think it would be a good idea to install anyways.