Don't install the default docker apt package
This is the docker package maintained by Ubuntu. However, we really
should be installing the official docker apt repository and then
installing it. For now, we just won't install it at all automatically.

Documentation on how to install it properly can be found at the below
link and also looking at my other projects (conjob) and seeing how it's
installed there.

https://docs.docker.com/engine/install/ubuntu/