Copy system files for HA server
The HA playbook doesn't depend on anything directly, so the task that
would normally do this wasn't.

I don't remember if there was a reason that I decided not to use
dependencies for the HA playbook and instead compose things. It might be
better to use deps instead, but for now we'll do it this way.