Increase test VM volume size
During install the test VM was running out of space. Likely we were
already close to using up all of the space and some changes to packages
put it over the edge.

It's also possibly indicative of a larger problem. However, after
increasing the capacity, the finaly used size is only slightly over the
old capacity (8GB) so it just seems like we were already close to
capacity before and slowly crept over the edge.

Additionally, this is meant to be run on a workstation which
realistically will have a much larger capacity.